+++ saleor/graphql/order/bulk_mutations/orders.py
@@ -1,9 +1,7 @@
 import graphene
 
-from saleor.extensions.manager import get_extensions_manager
-
 from ....order import events, models
-from ....order.utils import cancel_order
+from ....order.actions import cancel_order
 from ...core.mutations import BaseBulkMutation
 from ...core.types.common import OrderError
 from ..mutations.orders import clean_order_cancel
@@ -37,12 +35,8 @@ class OrderBulkCancel(BaseBulkMutation):
     @classmethod
     def bulk_action(cls, queryset, user, restock):
         for order in queryset:
-            cancel_order(user=user, order=order, restock=restock)
+            cancel_order(order=order, user=user, restock=restock)
             if restock:
                 events.fulfillment_restocked_items_event(
                     order=order, user=user, fulfillment=order
                 )
-            manager = get_extensions_manager()
-            manager.order_cancelled(order)
-            manager.order_updated(order)
-            events.order_canceled_event(order=order, user=user)
