+++ saleor/graphql/order/mutations/fulfillments.py
@@ -2,11 +2,14 @@ import graphene
 from django.core.exceptions import ValidationError
 from django.utils.translation import npgettext_lazy, pgettext_lazy
 
-from ....extensions.manager import get_extensions_manager
-from ....order import OrderStatus, events, models
-from ....order.emails import send_fulfillment_confirmation_to_customer
+from ....order import models
+from ....order.actions import (
+    cancel_fulfillment,
+    fulfillment_tracking_updated,
+    order_fulfilled,
+)
 from ....order.error_codes import OrderErrorCode
-from ....order.utils import cancel_fulfillment, fulfill_order_line, update_order_status
+from ....order.utils import fulfill_order_line
 from ...core.mutations import BaseMutation
 from ...core.types.common import OrderError
 from ...order.types import Fulfillment, Order
@@ -122,17 +125,12 @@ class FulfillmentCreate(BaseMutation):
             )
 
         fulfillment.lines.bulk_create(fulfillment_lines)
-        update_order_status(order)
-        events.fulfillment_fulfilled_items_event(
-            order=order, user=user, fulfillment_lines=fulfillment_lines
+        order_fulfilled(
+            fulfillment,
+            user,
+            fulfillment_lines,
+            cleaned_input.get("notify_customer", True),
         )
-        manager = get_extensions_manager()
-        manager.order_updated(order)
-        if order.status == OrderStatus.FULFILLED:
-            manager.order_fulfilled(order)
-
-        if cleaned_input.get("notify_customer", True):
-            send_fulfillment_confirmation_to_customer(order, fulfillment, user)
         return fulfillment
 
     @classmethod
@@ -172,13 +170,7 @@ class FulfillmentUpdateTracking(BaseMutation):
         fulfillment.tracking_number = tracking_number
         fulfillment.save()
         order = fulfillment.order
-        events.fulfillment_tracking_updated_event(
-            order=order,
-            user=info.context.user,
-            tracking_number=tracking_number,
-            fulfillment=fulfillment,
-        )
-        info.context.extensions.order_updated(order)
+        fulfillment_tracking_updated(fulfillment, info.context.user, tracking_number)
         return FulfillmentUpdateTracking(fulfillment=fulfillment, order=order)
 
 
@@ -218,6 +210,5 @@ class FulfillmentCancel(BaseMutation):
             )
 
         order = fulfillment.order
-        cancel_fulfillment(info.context.user, fulfillment, restock)
-        info.context.extensions.order_updated(order)
+        cancel_fulfillment(fulfillment, info.context.user, restock)
         return FulfillmentCancel(fulfillment=fulfillment, order=order)
