+++ saleor/product/utils/__init__.py
@@ -3,14 +3,9 @@ from urllib.parse import urlencode
 from django.conf import settings
 from django.db.models import F
 
-from ...checkout.utils import (
-    get_checkout_from_request,
-    get_or_create_checkout_from_request,
-)
 from ...core.taxes import TaxedMoney, zero_taxed_money
 from ...core.utils import get_paginator_items
 from ...core.utils.filters import get_now_sorted_by
-from ..forms import ProductForm
 from .availability import products_with_availability
 
 
@@ -61,22 +56,6 @@ def products_for_homepage(user, homepage_collection):
 def get_product_images(product):
     """Return list of product images that will be placed in product gallery."""
     return list(product.images.all())
-
-
-def handle_checkout_form(request, product, create_checkout=False):
-    if create_checkout:
-        checkout = get_or_create_checkout_from_request(request)
-    else:
-        checkout = get_checkout_from_request(request)
-    form = ProductForm(
-        checkout=checkout,
-        product=product,
-        data=request.POST or None,
-        discounts=request.discounts,
-        country=request.country,
-        extensions=request.extensions,
-    )
-    return form, checkout
 
 
 def products_for_checkout(user):
