+++ tests/api/test_fulfillment.py
@@ -35,10 +35,7 @@ CREATE_FULFILLMENT_QUERY = """
 """
 
 
-@patch(
-    "saleor.graphql.order.mutations.fulfillments."
-    "send_fulfillment_confirmation_to_customer"
-)
+@patch("saleor.order.actions.send_fulfillment_confirmation_to_customer")
 def test_create_fulfillment(
     mock_email_fulfillment,
     staff_api_client,
@@ -71,10 +68,7 @@ def test_create_fulfillment(
     assert mock_email_fulfillment.call_count == 1
 
 
-@patch(
-    "saleor.graphql.order.mutations.fulfillments."
-    "send_fulfillment_confirmation_to_customer"
-)
+@patch("saleor.order.actions.send_fulfillment_confirmation_to_customer")
 def test_create_fulfillment_with_empty_quantity(
     mock_send_fulfillment_confirmation,
     staff_api_client,
@@ -251,10 +245,7 @@ def test_cancel_fulfillment(
     assert event_cancel_fulfillment.user == staff_user
 
 
-@patch(
-    "saleor.graphql.order.mutations.fulfillments."
-    "send_fulfillment_confirmation_to_customer"
-)
+@patch("saleor.order.actions.send_fulfillment_confirmation_to_customer")
 def test_create_digital_fulfillment(
     mock_email_fulfillment,
     digital_content,
